Extract shared helpers for pipeline tests
